Add external facts support

+++ src/js/renderer/causal-view/render/graph-manager.js
@@ -26,47 +26,37 @@ export class GraphManager {
       : causeId
   }
 
-  #setStratify() {
-    this.stratify = d3dag
-      .graphStratify()
-      .id((nodeData) => CausalViewNodeUtils.getNodeId(nodeData))
-      .parentIds(
-        (nodeData) => {
-          const causeIds = CausalModelUtils.getCausesIdsUnique(nodeData);
-          const parentIdsInGraph = this.#toParentIdsInGraph(causeIds);
-          return parentIdsInGraph;
-        }
-      );
-  }
-
-  #toParentIdsInGraph(causeIds) {
-    return causeIds.map(causeId => this.getParentIdInGraph(causeId));
-  }
-
   // Sets directed acyclic graph by causal model facts
   resetGraph(nodesData) {
+    nodesData = [...nodesData, ...CausalViewNodeUtils.getExternalCauseNodes(nodesData)];
     this.#resetRelatedData(nodesData);
     this.mutGraph = this.stratify(nodesData);
   }
 
-  #resetRelatedData(nodesData) {
-    this.#blockIdsByMappedCauseIds = new Map();
-    for (const nodeData of nodesData) {
-      if (nodeData.block) {
-        for (const mappedCauseId of BlockUtils.getCauseIdsMappedAsConsequences(
-          nodeData.block)) {
-          this.#blockIdsByMappedCauseIds.set(mappedCauseId, nodeData.block.id);
-        }
+  ensureNodeDataPositionsSet() {
+    for (const nodeData of this.getNodesData()) {
+      if (!nodeData.x || !nodeData.y) {
+        const node = this.getNodeById(CausalViewNodeUtils.getNodeId(nodeData));
+        nodeData.x = node.x;
+        nodeData.y = node.y;
       }
     }
   }
 
-  addNodeWithData(nodeData) {
-    const newNode = this.mutGraph.node(nodeData);
+  setNodeWithData(nodeData) {
+    let node = this.tryGetNodeById(CausalViewNodeUtils.getNodeId(nodeData));
+    if (!node) {
+      node = this.mutGraph.node(nodeData);
+    } else {
+      node.data = nodeData;
+    }
 
-    newNode.ux = nodeData.x;
-    newNode.uy = nodeData.y;
-    return newNode;
+    if (!node.ux || !node.uy) {
+      node.ux = nodeData.x;
+      node.uy = nodeData.y;
+    }
+
+    return node;
   }
 
   removeNode(nodeId) {
@@ -75,7 +65,25 @@ export class GraphManager {
       console.error("Node to remove is not found. ", nodeToRemove);
       return;
     }
-    nodeToRemove.delete();
+
+    if (nodeToRemove.isExternal) {
+      console.error("External cause cannot be removed");
+      return;
+    }
+
+    if (CausalViewNodeUtils.isReferencedBySomeNode(this.getNodesData(), nodeId)) {
+      const externalCauseId = CausalViewNodeUtils.getNodeId(nodeToRemove.data);
+      nodeToRemove.data.id = externalCauseId;
+      nodeToRemove.data.isExternal = true;
+      delete nodeToRemove.data.fact;
+      delete nodeToRemove.data.block;
+
+      for (const link of this.getLinksByTarget(externalCauseId)) {
+        link.delete();
+      }
+    } else {
+      nodeToRemove.delete();
+    }
   }
 
   addLink(sourceId, targetId) {
@@ -93,6 +101,17 @@ export class GraphManager {
         // Old implementation that doesn't support blocks:
         // CausalViewNodeUtils.getNodeId(link.source.data) === sourceId &&
         CausalViewNodeUtils.getNodeId(link.target.data) === targetId
+    );
+    return result;
+  }
+
+  /**
+   * Gets edges by the node that has them
+   */
+  getLinksByTarget(targetNodeId) {
+    const result = Array.from(this.mutGraph.links()).filter(
+      (link) =>
+        CausalViewNodeUtils.getNodeId(link.target.data) === targetNodeId
     );
     return result;
   }
@@ -117,16 +136,49 @@ export class GraphManager {
   }
 
   getNodeById(nodeId) {
-    const node = this.getNodes().find(
-      (node) => CausalViewNodeUtils.canBeReferencedInViewBy(node.data, nodeId)
-    );
+    const node = this.tryGetNodeById(nodeId);
     if (!node) {
       console.error("Node not found by id: ", nodeId);
     }
     return node;
   }
 
+  tryGetNodeById(nodeId) {
+    return this.getNodes().find(
+      (node) => CausalViewNodeUtils.canBeReferencedInViewBy(node.data, nodeId)
+    );
+  }
+
   getNodeDataById(nodeId) {
     return this.getNodeById(nodeId).data;
+  }
+
+  #resetRelatedData(nodesData) {
+    this.#blockIdsByMappedCauseIds = new Map();
+    for (const nodeData of nodesData) {
+      if (nodeData.block) {
+        for (const mappedCauseId of BlockUtils.getCauseIdsMappedAsConsequences(
+          nodeData.block)) {
+          this.#blockIdsByMappedCauseIds.set(mappedCauseId, nodeData.block.id);
+        }
+      }
+    }
+  }
+
+  #setStratify() {
+    this.stratify = d3dag
+      .graphStratify()
+      .id((nodeData) => CausalViewNodeUtils.getNodeId(nodeData))
+      .parentIds(
+        (nodeData) => {
+          const causeIds = CausalModelUtils.getCausesIdsUnique(nodeData);
+          const parentIdsInGraph = this.#toParentIdsInGraph(causeIds);
+          return parentIdsInGraph;
+        }
+      );
+  }
+
+  #toParentIdsInGraph(causeIds) {
+    return causeIds.map(causeId => this.getParentIdInGraph(causeId));
   }
 }
